[CXF-3308] Change to using long instead of BigInteger for WS-RM sequences to make the code easier to maintain and use. Patch from Dennis Sosnoski applied
git-svn-id: https://svn.apache.org/repos/asf/cxf/trunk@1070579 13f79535-47bb-0310-9956-ffa450edef68

+++ distribution/src/main/release/samples/ws_rm/src/demo/ws_rm/common/MessageLossSimulator.java
@@ -21,7 +21,6 @@ package demo.ws_rm.common;
 
 import java.io.IOException;
 import java.io.OutputStream;
-import java.math.BigInteger;
 import java.util.ListIterator;
 import java.util.logging.Level;
 import java.util.logging.Logger;
@@ -107,7 +106,7 @@ public class MessageLossSimulator extends AbstractPhaseInterceptor<Message> {
         @Override
         protected void onFirstWrite() throws IOException {
             if (LOG.isLoggable(Level.FINE)) {
-                BigInteger nr = RMContextUtils.retrieveRMProperties(outMessage, true)
+                Long nr = RMContextUtils.retrieveRMProperties(outMessage, true)
                     .getSequence().getMessageNumber();
                 LOG.fine("Losing message " + nr);
             }
